Retrieve csaf vulns
