+++ mirror-service/src/main/java/org/dependencytrack/vulnmirror/datasource/csaf/CsafMirrorState.java
@@ -18,25 +18,8 @@
  */
 package org.dependencytrack.vulnmirror.datasource.csaf;
 
-import java.util.Optional;
+import io.quarkus.runtime.annotations.RegisterForReflection;
 
-import org.eclipse.microprofile.config.inject.ConfigProperty;
-
-import jakarta.enterprise.context.ApplicationScoped;
-import jakarta.inject.Provider;
-
-@ApplicationScoped
-public class CSAFConfig {
-
-    private final Provider<Optional<Boolean>> enabledProvider;
-
-    CSAFConfig(
-        @ConfigProperty(name = "dtrack.vuln-source.csaf.enabled") final Provider<Optional<Boolean>> enabledProvider
-    ) {
-        this.enabledProvider = enabledProvider;
-    }
-
-    Optional<Boolean> enabled() {
-        return enabledProvider.get();
-    }
+@RegisterForReflection
+public record CsafMirrorState(long lastModifiedEpochSeconds) {
 }
