+++ pom.xml
@@ -95,7 +95,7 @@
     <quarkus.wiremock.version>1.3.3</quarkus.wiremock.version>
     <lib.brotli.version>0.1.2</lib.brotli.version>
     <lib.versatile.version>0.7.0</lib.versatile.version>
-    <lib.kotlin-csaf.version>0.0.1</lib.kotlin-csaf.version>
+    <lib.csaf-retrieval.version>0.0.2</lib.csaf-retrieval.version>
 
     <!-- Plugin Versions -->
     <plugin.jacoco.version>0.8.13</plugin.jacoco.version>
@@ -386,8 +386,8 @@
       
       <dependency>
         <groupId>io.github.csaf-sbom</groupId>
-        <artifactId>csaf-import</artifactId>
-        <version>${lib.kotlin-csaf.version}</version>        
+        <artifactId>csaf-retrieval</artifactId>
+        <version>${lib.csaf-retrieval.version}</version>        
       </dependency>
     </dependencies>
   </dependencyManagement>
